Import v0.3.0
This imports version 0.3.0 of the profiler from the private repository
it used while under development. It is missing some of the supporting
files like docker containers, CI infra, etc to reduce the amount of
scope for security folks to audit for going public.

Aside from these omissions, the only other change is the README. It has
been rewritten to better serve its new audience.