Link libc dynamically
On x86_64-alpine-linux-musl, if libc is linked in statically,
accidentally or otherwise, then any symbol used by VDSO is likely to
cause a sigsegv when used, because on this platform __vdsosym cannot
open a private symbol. The symbol cannot be public, because then it
would open up symbol conflict opportunities with things which link
against libc (aka almost everything!).

So, always dynamically link against libc.