Re-initialize static variables on startup/minit (#41)
* Reinitialize static variables on startup/minit

This is required because when Apache does a reload, the processes will
do a shutdown sequence but keep the process alive. It then goes through
startup again. Without re-initializing statics in startup they may be
stale and cause issues such as crashes.

Note that uv_once and pthread_once don't provide the necessary routines
to re-initialize a once, so I implemented a once component which allows
it, with restrictions the programmer will have to follow, of course.

* Ensure module_entry's .handle is NULL