Support another envvar for CPU time; fix memory leak (#43)
The profiler recognizes `DD_PROFILING_EXPERIMENTAL_CPU_ENABLED`, but
the feature is documented as
`DD_PROFILING_EXPERIMENTAL_CPU_TIME_ENABLED`. This adds support for the
latter and prefers it.

Also fixes a memory leak which occurs in env var handling. Given this
happens only once per process, the leak is not very significant.